Move struct CacheVol into Stripe.h
